// src/modules/garmentsProduction/reports/finishing/garments-dispatch-details-report/garments-dispatch-details-report-type.ts
export interface IGarmentsDispatchDetailsReportDto {
  REJECTDATE: Date;
  REPORT_NO: string;
  BUYERNAME: string;
  STYLENO: string;
  ITEMTYPE: string;
  PONO: string;
  COLORNAME: string;
  SIZENAME: string;
  REJECTQTY: number;
  REJECTNAME: string;
  SEND_TO: string;
  REMARKS: string;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  COMPANY_REMARKS: string;
}